Improve phoenix cleaning

+++ bash_scripts/phoenix_temp_clean.sh
@@ -14,8 +14,8 @@ show_help(){
     echo  "Apache Phoenix Cleaner Utility"
     echo  " *****************************************************************************"
     echo  " *****************************************************************************"
-    echo  "Usage: $0 [days]" 
-    echo  "          Search and deletes Apache Phoenix temp files older than [days] days."
+    echo  "Usage: $0 [minutes]" 
+    echo  "          Search and deletes Apache Phoenix temp files older than [minutes] minutes."
     echo  "          See PHOENIX-4910 and PHOENIX-2850 to understand why is this necessary."    
     echo  "                                                                           "  
     echo  "          Temp directory is defined by _JAVA_OPTIONS environment variable,"  
@@ -40,9 +40,14 @@ fi
 if cd ${JAVA_IO_TMPDIR}; then
     for file in $(find . -regextype posix-extended -regex './[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{31}\.tmp' -user hbase -mmin +$1); do 
         echo $file
-        rm $file
-        COUNTER=$(($(cat $TEMPFILE) + 1))
-        echo $COUNTER > $TEMPFILE
+        # dirty check if file is in use
+        if [[ $(lsof $file) ]]; then
+            echo "This file seem to be currently in use"
+        else
+            rm $file
+            COUNTER=$(($(cat $TEMPFILE) + 1))
+            echo $COUNTER > $TEMPFILE
+        fi
     done
 fi
 log "Deleted $(cat $TEMPFILE) files."
